% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTE.R
\name{Select.HTGs}
\alias{Select.HTGs}
\title{Select HTGs under a group variable.}
\usage{
Select.HTGs(GTE, bins = 0.1, gte.ratio = 0.95)
}
\arguments{
\item{GTE}{GTE result.}

\item{bins}{Bins.}

\item{gte.ratio}{Ratio of selected HTGs to overall GTE.}
}
\description{
Select HTGs under a group variable.
}
