% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTE.R
\name{select_hbgs}
\alias{select_hbgs}
\title{Select HBGs using GTE vector.}
\usage{
select_hbgs(gte, bins = 0.1, gte.ratio = 0.95, is.sort = T)
}
\arguments{
\item{gte}{Named GTE vector.}

\item{bins}{Bins.}

\item{gte.ratio}{Ratio of selected HBGs to overall GTE.}

\item{is.sort}{Whether to sort genes by GTE from largest to smallest.}
}
\description{
Select HBGs using GTE vector.
}
