% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEs.R
\name{select_htgs}
\alias{select_htgs}
\title{Select HTGs using GTEs vector.}
\usage{
select_htgs(gtes, bins = 0.1, gte.ratio = 0.95, is.sort = T)
}
\arguments{
\item{gtes}{Named GTEs vector.}

\item{bins}{Bins.}

\item{gte.ratio}{Ratio of selected HTGs to overall GTEs.}

\item{is.sort}{Whether to sort genes by GTEs from largest to smallest.}
}
\description{
Select HTGs using GTEs vector.
}
