% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEs.R
\name{Run.GroupTechEffects}
\alias{Run.GroupTechEffects}
\title{Compute the group technical effects.}
\usage{
Run.GroupTechEffects(X, meta, g_factor, b_factor, do.scale = F)
}
\arguments{
\item{X}{Input data matrix.}

\item{meta}{Input metadata (data.frame).}

\item{g_factor}{Group variable (s).}

\item{b_factor}{Batch variable (s).}

\item{do.scale}{Whether to perform scaling.}
}
\description{
Compute the group technical effects.
}
