% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEs.R
\name{Select.HTGs}
\alias{Select.HTGs}
\title{Select HTGs under a group variable.}
\usage{
Select.HTGs(GTEs, bins = 0.1, gte.ratio = 0.95)
}
\arguments{
\item{GTEs}{GTEs object.}

\item{bins}{Bins.}

\item{gte.ratio}{Ratio of selected HTGs to overall GTEs.}
}
\description{
Select HTGs under a group variable.
}
