% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTE.R
\name{select_htgs}
\alias{select_htgs}
\title{Select HTGs using GTE vector.}
\usage{
select_htgs(gte, bins = 0.1, gte.ratio = 0.95, is.sort = T)
}
\arguments{
\item{gte}{Named GTE vector.}

\item{bins}{Bins.}

\item{gte.ratio}{Ratio of selected HTGs to overall GTE.}

\item{is.sort}{Whether to sort genes by GTE from largest to smallest.}
}
\description{
Select HTGs using GTE vector.
}
